+++ ErsatzTV.Application/ProgramSchedules/Commands/UpdateProgramSchedule.cs
@@ -1,4 +1,5 @@
 ﻿using ErsatzTV.Core;
+using ErsatzTV.Core.Scheduling;
 
 namespace ErsatzTV.Application.ProgramSchedules;
 
@@ -8,4 +9,5 @@ public record UpdateProgramSchedule(
     bool KeepMultiPartEpisodesTogether,
     bool TreatCollectionsAsShows,
     bool ShuffleScheduleItems,
-    bool RandomStartPoint) : IRequest<Either<BaseError, UpdateProgramScheduleResult>>;
+    bool RandomStartPoint,
+    FixedStartTimeBehavior FixedStartTimeBehavior) : IRequest<Either<BaseError, UpdateProgramScheduleResult>>;