+++ ErsatzTV.Core/Scheduling/PlayoutModeSchedulerBase.cs
@@ -71,7 +71,21 @@ public abstract class PlayoutModeSchedulerBase<T> : IPlayoutModeScheduler<T> whe
             //     itemStartTime.TotalMilliseconds);
 
             // need to wrap to the next day if appropriate
-            startTime = startTime.TimeOfDay > itemStartTime ? result.AddDays(1) : result;
+            FixedStartTimeBehavior? fixedStartTimeBehavior = scheduleItem.FixedStartTimeBehavior;
+            if (fixedStartTimeBehavior is null && scheduleItem.ProgramSchedule is not null)
+                fixedStartTimeBehavior = scheduleItem.ProgramSchedule.FixedStartTimeBehavior;
+            switch (fixedStartTimeBehavior)
+            {
+                case FixedStartTimeBehavior.Flexible:
+                    // only wait for times on the same day
+                    if (result.Day == startTime.Day && result.TimeOfDay > startTime.TimeOfDay)
+                        startTime = result;
+                    break;
+                case FixedStartTimeBehavior.Strict:
+                default:
+                    startTime = startTime.TimeOfDay > itemStartTime ? result.AddDays(1) : result;
+                    break;
+            }
         }
 
         return startTime;