+++ ErsatzTV.Infrastructure.MySql/Migrations/20250530085651_Add_ProgramScheduleItem_FixedStartTimeBehavior.cs
@@ -1,0 +1,28 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace ErsatzTV.Infrastructure.MySql.Migrations
+{
+    /// <inheritdoc />
+    public partial class Add_ProgramScheduleItem_FixedStartTimeBehavior : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<int>(
+                name: "FixedStartTimeBehavior",
+                table: "ProgramScheduleItem",
+                type: "int",
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "FixedStartTimeBehavior",
+                table: "ProgramScheduleItem");
+        }
+    }
+}