+++ ErsatzTV.Infrastructure.Sqlite/Migrations/20250530083219_Add_ProgramSchedule_FixedStartTimeBehavior.cs
@@ -1,0 +1,29 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace ErsatzTV.Infrastructure.Sqlite.Migrations
+{
+    /// <inheritdoc />
+    public partial class Add_ProgramSchedule_FixedStartTimeBehavior : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<int>(
+                name: "FixedStartTimeBehavior",
+                table: "ProgramSchedule",
+                type: "INTEGER",
+                nullable: false,
+                defaultValue: 0);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "FixedStartTimeBehavior",
+                table: "ProgramSchedule");
+        }
+    }
+}